Rename bigNumberMath to bnMath
