+++ test/integration/StakingProducts/recalculateEffectiveWeightsForAllProducts.js
@@ -6,7 +6,7 @@ const { calculateFirstTrancheId } = require('../utils/staking');
 const { daysToSeconds } = require('../../../lib/helpers');
 const { buyCover, ETH_ASSET_ID } = require('../utils/cover');
 const { setNextBlockTime } = require('../utils').evm;
-const { divCeil } = require('../utils').bigNumberMath;
+const { divCeil } = require('../utils').bnMath;
 const { getInternalPrice } = require('../../utils/rammCalculations');
 const { roundUpToNearestAllocationUnit } = require('../../unit/StakingPool/helpers');
 const { loadFixture } = require('@nomicfoundation/hardhat-network-helpers');
